fix: resolve all 130 clippy/rustc warnings workspace-wide
- clippy --fix mechanical batch (clone-on-Copy, useless vec!, map_or,
  same-type casts, unused imports/vars, manual ASCII case compare, ...)
- doc list continuation indents normalized in 75 doc comments
- deprecated brackets_2026 test callers -> brackets_for(2026, ...)
- if let Err(_) panic guards -> .is_err() (live_ticks workers)
- guard clauses rewritten conjunctively (double_barrier, risk_of_ruin,
  seagull_spread, crypto book_depth) - NaN rejection preserved
- range loops -> iterator forms (debt_avalanche, debt_snowball,
  strategy_calculators fixture, Nelder-Mead shrink in optimize)
- digest test uses struct literal with ..Default::default()

+++ crates/traderview-core/src/black_scholes.rs
@@ -42,7 +42,7 @@ mod tests {
     fn put_call_parity_holds() {
         let (s, k, t, r, q, sig) = (100.0, 95.0, 0.5, 0.04, 0.01, 0.3);
         let lhs = call(s, k, t, r, q, sig) - put(s, k, t, r, q, sig);
-        let rhs = s * (-q * t as f64).exp() - k * (-r * t as f64).exp();
+        let rhs = s * (-q * t).exp() - k * (-r * t).exp();
         assert!((lhs - rhs).abs() < 1e-12, "{lhs} vs {rhs}");
     }
